# swagger.yaml
openapi: 3.0.0
info:
  title: Order API
  description: API for a service managing orders.
  version: 0.1.9
host: api.example.com
schemes:
  - https
paths:
  /orders:
    post:
      summary: Add a new order
      requestBody:
        description: Add a new order
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                products:
                  type: string
                amount: 
                  type: string
      responses:
        '201':
          description: Created
    get:
      summary: Returns the order list.
      parameters:
        - in: path
          name: orders
          schema:
            type : string
          required: true
          description: ID of the order to get
      description: Return the whole order list.
      responses:
        '200': 
          description: OK
  /orders/{orderId}:
    get:
      summary: Returns a single order.
      parameters:
        - in: path
          name: orderId
          schema:
            type : string
          required: true
          description: ID of the order to get
      description: Return a single order based on id.
      responses:
        '200': 
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties: 
                  id:
                    type: string
                    description: The order ID.
                    example: XXXX_1234
                  products:
                    type: object
                    description: The products in order.
                    example: "T-Shirt"
                  amount:
                    type: integer
                    description: The amount of all products together.
                    example: 10
    delete:
      summary: "Deletes an order."
      description: "Delete a single order based on id."
      parameters:
      - name: "orderId"
        in: "path"
        description: "Order id to delete"
        required: true
        type: "string"
        example: "XXXX_1234"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
    post:
      summary: "Updates an order in the order list."
      description: "Updates an order in the order list."
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of order that needs to be updated"
        required: true
        type: "string"
        example: "XXXX_1234"
      - name: "amount"
        in: "path"
        description: "Updated amount of the order"
        type: integer
      responses:
        "405":
          description: "Invalid input"